refactor: email modify code / link hooks naming
