+++ src/components/Forms/ConfirmEmailLinkForm/hooks/useConfirmEmailLink.ts
@@ -5,10 +5,7 @@ import {useNavigation} from '@react-navigation/native';
 import {NativeStackNavigationProp} from '@react-navigation/native-stack';
 import {AccountActions} from '@store/modules/Account/actions';
 import {failedReasonSelector} from '@store/modules/UtilityProcessStatuses/selectors';
-import {
-  emailVerificationCodeSelector,
-  temporaryEmailSelector,
-} from '@store/modules/Validation/selectors';
+import {temporaryEmailSelector} from '@store/modules/Validation/selectors';
 import {showError} from '@utils/errors';
 import {useEffect} from 'react';
 import {useDispatch, useSelector} from 'react-redux';
@@ -18,14 +15,13 @@ export const useConfirmEmailLink = () => {
   const navigation =
     useNavigation<NativeStackNavigationProp<AuthStackParamList>>();
   const email = useSelector(temporaryEmailSelector, () => true);
-  const code = useSelector(emailVerificationCodeSelector, () => true);
 
   const validateError = useSelector(
-    failedReasonSelector.bind(null, AccountActions.MODIFY_EMAIL_WITH_CODE),
+    failedReasonSelector.bind(null, AccountActions.MODIFY_EMAIL_WITH_LINK),
   );
 
   const goBack = () => {
-    dispatch(AccountActions.MODIFY_EMAIL_WITH_CODE.RESET.create());
+    dispatch(AccountActions.MODIFY_EMAIL_WITH_LINK.RESET.create());
   };
 
   useEffect(() => {
@@ -36,14 +32,13 @@ export const useConfirmEmailLink = () => {
 
   useEffect(
     () => () => {
-      dispatch(AccountActions.MODIFY_EMAIL_WITH_CODE.RESET.create());
+      dispatch(AccountActions.MODIFY_EMAIL_WITH_LINK.RESET.create());
     },
     [dispatch],
   );
 
   return {
     email,
-    code,
     goBack,
   };
 };
